pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Stream to Frame Service"
authors = ["Lukas Fam <takeruadelbert@gmail.com>"]
license = "PT. Surya Teknologi Nasional"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "~=3.7.2"
aiojobs = "~=0.2.2"
kafka-python = "~=2.0.2"
opencv-python = "~=4.4.0.46"
PyMySQL = "~=0.10.1"
arrow = "~=0.17.0"
cryptography = "~=3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
